reorg template types
